% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpb_series_data.R
\docType{data}
\name{wpb_series_data}
\alias{wpb_series_data}
\title{Time series data on global prison population.}
\format{
A data frame with 1567 rows and 4 variables.
}
\usage{
wpb_series_data
}
\description{
A dataset containing information on prison population rates and total population over several years.
}
\details{
\itemize{
  \item country \code{character}.
  \item year \code{numeric}. 
  \item prison_population_total \code{numeric}.
  \item prison_population_rate \code{numeric}.
}
}
\keyword{datasets}
