% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpb_tables_data.R
\docType{data}
\name{wpb_tables_data}
\alias{wpb_tables_data}
\title{Data on global prison population.}
\format{
A data frame with 226 rows and 11 variables.
}
\usage{
wpb_tables_data
}
\description{
A dataset containing information on prison population for all countries available.
}
\details{
\itemize{
  \item country \code{character}.
  \item prison_population_total \code{integer}.
  \item prison_population_rate \code{integer}.
  \item female_prisoners \code{numeric}.
  \item pre_trial_detainees \code{numeric}.
  \item foreign_prisoners \code{numeric}.
  \item occupancy_level \code{numeric}.
  \item juvenile_prisoners \code{numeric}.
  \item number_institutions \code{numeric}.
  \item iso_a2 \code{character}.
  \item geometry \code{sfc_MULTIPOLYGON}.
}
}
\keyword{datasets}
