% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpb_tables_data.R
\docType{data}
\name{wpb_tables_data}
\alias{wpb_tables_data}
\title{Data on global prison population.}
\format{A data frame with 226 rows and 11 variables.}
\usage{
wpb_tables_data
}
\description{
A dataset containing information on prison population for all countries available.
}
\keyword{datasets}
